Organize classes into suitable packages and update runtest.bat script
Previously, all classes were stored in the same folder, making the
project harder to manage and extend. There was no clear separation of
concerns, and the runtest.bat script did not account for any package
structure, leading to build issues.

Changes made:
* Organized classes into appropriate packages:
  - Moved Aether and DukeException to the 'aether' package.
  - Moved Task, Todo, Deadline, and Event classes to 'aether.task' package.
  - Moved Display class to 'aether.ui' package.
* Updated package declarations in all classes to reflect the new structure.
* Updated runtest.bat:
  - Modified compilation to include 'aether', 'aether.task', and 'aether.ui'.
  - Adjusted execution command to run the 'aether.Aether' main class.
  - Ensured proper comparison of test outputs (ACTUAL.TXT vs EXPECTED.TXT).

These changes improve the organization of the codebase, making it more
maintainable and easier to extend in the future, while ensuring the
test script is aligned with the new package structure.
