+++ src/main/java/aether/task/Deadline.java
@@ -1,0 +1,18 @@
+package aether.task;
+
+/**
+ * Deadline class for tasks with deadlines.
+ */
+public class Deadline extends Task {
+    protected String by;
+
+    public Deadline(String description, String by) {
+        super(description);
+        this.by = by;
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + by + ")";
+    }
+}
