+++ src/main/java/aether/task/Event.java
@@ -1,0 +1,20 @@
+package aether.task;
+
+/**
+ * Event class for tasks with start and end times.
+ */
+public class Event extends Task {
+    protected String from;
+    protected String to;
+
+    public Event(String description, String from, String to) {
+        super(description);
+        this.from = from;
+        this.to = to;
+    }
+
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + " (from: " + from + " to: " + to + ")";
+    }
+}
