+++ src/main/java/aether/task/Task.java
@@ -1,0 +1,27 @@
+package aether.task;
+
+/**
+ * Task class, the base class for all types of tasks.
+ */
+public class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public void setDone(boolean isDone) {
+        this.isDone = isDone;
+    }
+
+    public String getStatus() {
+        return (isDone ? "✓" : " ");
+    }
+
+    @Override
+    public String toString() {
+        return "[" + getStatus() + "] " + description;
+    }
+}
