+++ src/main/java/aether/task/Todo.java
@@ -1,0 +1,15 @@
+package aether.task;
+
+/**
+ * Todo class for simple todo tasks.
+ */
+public class Todo extends Task {
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
